test(api-core): move site-explorer switch tests from api-core
Move switch-related site explorer tests into the site-explorer
crate. Adapt the tests to use the site-explorer test harness
environment.
